replace hoverProvider.ts
